% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{CompositeFilter}
\alias{CompositeFilter}
\title{CompositeFilter Object}
\usage{
CompositeFilter(op = NULL, filters = NULL)
}
\arguments{
\item{op}{The operator for combining multiple filters}

\item{filters}{The list of filters to combine}
}
\value{
CompositeFilter object
}
\description{
CompositeFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter that merges multiple other filters using the given operator.
}
