% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Entity.properties}
\alias{Entity.properties}
\title{Entity.properties Object}
\usage{
Entity.properties()
}
\value{
Entity.properties object
}
\description{
Entity.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The entity's properties.The map's keys are property names.A property name matching regex `__.*__` is reserved.A reserved property name is forbidden in certain documented contexts.The name must not contain more than 500 characters.The name cannot be `''`.
}
\seealso{
Other Entity functions: \code{\link{Entity}}
}

