% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Entity}
\alias{Entity}
\title{Entity Object}
\usage{
Entity(Entity.properties = NULL, properties = NULL, key = NULL)
}
\arguments{
\item{Entity.properties}{The \link{Entity.properties} object or list of objects}

\item{properties}{The entity's properties}

\item{key}{The entity's key}
}
\value{
Entity object
}
\description{
Entity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Datastore data object.An entity is limited to 1 megabyte when stored. That _roughly_corresponds to a limit of 1 megabyte for the serialized form of thismessage.
}
\seealso{
Other Entity functions: \code{\link{Entity.properties}}
}

