% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Key}
\alias{Key}
\title{Key Object}
\usage{
Key(partitionId = NULL, path = NULL)
}
\arguments{
\item{partitionId}{Entities are partitioned into subsets, currently identified by a project}

\item{path}{The entity path}
}
\value{
Key object
}
\description{
Key Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A unique identifier for an entity.If a key's partition ID or any of its path kinds or names arereserved/read-only, the key is reserved/read-only.A reserved/read-only key is forbidden in certain documented contexts.
}
