% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{LatLng}
\alias{LatLng}
\title{LatLng Object}
\usage{
LatLng(latitude = NULL, longitude = NULL)
}
\arguments{
\item{latitude}{The latitude in degrees}

\item{longitude}{The longitude in degrees}
}
\value{
LatLng object
}
\description{
LatLng Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object representing a latitude/longitude pair. This is expressed as a pairof doubles representing degrees latitude and degrees longitude. Unlessspecified otherwise, this must conform to the<a href='http://www.unoosa.org/pdf/icg/2012/template/WGS_84.pdf'>WGS84standard</a>. Values must be within normalized ranges.Example of normalization code in Python:    def NormalizeLongitude(longitude):      '''Wraps decimal degrees longitude to [-180.0, 180.0].'''      q, r = divmod(longitude, 360.0)      if r > 180.0 or (r == 180.0 and q <= -1.0):        return r - 360.0      return r    def NormalizeLatLng(latitude, longitude):      '''Wraps decimal degrees latitude and longitude to      [-90.0, 90.0] and [-180.0, 180.0], respectively.'''      r = latitude % 360.0      if r <= 90.0:        return r, NormalizeLongitude(longitude)      elif r >= 270.0:        return r - 360, NormalizeLongitude(longitude)      else:        return 180 - r, NormalizeLongitude(longitude + 180.0)    assert 180.0 == NormalizeLongitude(180.0)    assert -180.0 == NormalizeLongitude(-180.0)    assert -179.0 == NormalizeLongitude(181.0)    assert (0.0, 0.0) == NormalizeLatLng(360.0, 0.0)    assert (0.0, 0.0) == NormalizeLatLng(-360.0, 0.0)    assert (85.0, 180.0) == NormalizeLatLng(95.0, 0.0)    assert (-85.0, -170.0) == NormalizeLatLng(-95.0, 10.0)    assert (90.0, 10.0) == NormalizeLatLng(90.0, 10.0)    assert (-90.0, -10.0) == NormalizeLatLng(-90.0, -10.0)    assert (0.0, -170.0) == NormalizeLatLng(-180.0, 10.0)    assert (0.0, -170.0) == NormalizeLatLng(180.0, 10.0)    assert (-90.0, 10.0) == NormalizeLatLng(270.0, 10.0)    assert (90.0, 10.0) == NormalizeLatLng(-270.0, 10.0)The code in logs/storage/validator/logs_validator_traits.cc treats this typeas if it were annotated as ST_LOCATION.
}
