% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PartitionId}
\alias{PartitionId}
\title{PartitionId Object}
\usage{
PartitionId(namespaceId = NULL, projectId = NULL)
}
\arguments{
\item{namespaceId}{If not empty, the ID of the namespace to which the entities belong}

\item{projectId}{The ID of the project to which the entities belong}
}
\value{
PartitionId object
}
\description{
PartitionId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A partition ID identifies a grouping of entities. The grouping is alwaysby project and namespace, however the namespace ID may be empty.A partition ID contains several dimensions:project ID and namespace ID.Partition dimensions:- May be `''`.- Must be valid UTF-8 bytes.- Must have values that match regex `[A-Za-z\d\.\-_]{1,100}`If the value of any dimension matches regex `__.*__`, the partition isreserved/read-only.A reserved/read-only partition ID is forbidden in certain documentedcontexts.Foreign partition IDs (in which the project ID doesnot match the context project ID ) are discouraged.Reads and writes of foreign partition IDs may fail if the project is not in an active state.
}
