% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{Projection}
\alias{Projection}
\title{Projection Object}
\usage{
Projection(property = NULL)
}
\arguments{
\item{property}{The property to project}
}
\value{
Projection object
}
\description{
Projection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a property in a projection.
}
