% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_functions.R
\name{projects.allocateIds}
\alias{projects.allocateIds}
\title{Allocates IDs for the given keys, which is useful for referencing an entitybefore it is inserted.}
\usage{
projects.allocateIds(AllocateIdsRequest, projectId)
}
\arguments{
\item{AllocateIdsRequest}{The \link{AllocateIdsRequest} object to pass to this method}

\item{projectId}{The ID of the project against which to make the request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/datastore/}{Google Documentation}

Other AllocateIdsRequest functions: \code{\link{AllocateIdsRequest}}
}

