% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PropertyOrder}
\alias{PropertyOrder}
\title{PropertyOrder Object}
\usage{
PropertyOrder(direction = NULL, property = NULL)
}
\arguments{
\item{direction}{The direction to order by}

\item{property}{The property to order by}
}
\value{
PropertyOrder object
}
\description{
PropertyOrder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The desired order for a specific property.
}
