% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{PropertyReference}
\alias{PropertyReference}
\title{PropertyReference Object}
\usage{
PropertyReference(name = NULL)
}
\arguments{
\item{name}{The name of the property}
}
\value{
PropertyReference object
}
\description{
PropertyReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a property relative to the kind expressions.
}

