% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_objects.R
\name{RollbackResponse}
\alias{RollbackResponse}
\title{RollbackResponse Object}
\usage{
RollbackResponse()
}
\value{
RollbackResponse object
}
\description{
RollbackResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Datastore.Rollback.(an empty message).
}
